Added the subscription and renamed publisher to mqttClient.
